+++ CHANGELOG.md
@@ -50,6 +50,11 @@ Breaking changes
   number.
 - @iov/jsonrpc: Remove `parseJsonRpcResponse` and `parseJsonRpcError`.
 
+## 0.14.3
+
+- @iov/socket: Fix bug in `SocketWrapper` where the timeout error was triggered
+  even if the connection was successfully established.
+
 ## 0.14.2
 
 - @iov/jsonrpc: Add `makeJsonRpcId`.
@@ -62,6 +67,11 @@ Breaking changes
   `SigningServerCore.signAndPost`, `JsonRpcSigningServer.handleUnchecked` and
   `JsonRpcSigningServer.handleChecked` which allows passing request meta
   information from the request handler to the authorization callbacks.
+
+## 0.13.8
+
+- @iov/socket: Fix bug in `SocketWrapper` where the timeout error was triggered
+  even if the connection was successfully established.
 
 ## 0.13.7